[Feat] member, instructor 구조 정의
