+++ app/constants.py
@@ -12,3 +12,8 @@ NO_DUPLICATION = -3
 USER_TYPE_MEMBER = 'm'  
 ## Instructor    
 USER_TYPE_INSTRUCTOR = 'n'
+
+##file_path
+MEMBER_PATH = 'data/members.csv'
+INSTRUCTOR_PATH = 'data/instructor.csv'
+CLASS_PATH = 'data/members.csv'
