+++ models.py
@@ -1,0 +1,13 @@
+#회원 정보 클래스 생성
+class Member_Data:
+    id : str
+    pw : str
+    name : str
+    ph : str
+    tnum : int
+
+class Instructor_Data:
+    id : str
+    pw : str
+    name : str
+    ph : str
